Dataset processing is now done with COCO
Not anymore to mmdetection's middle format

+++ process_dataset/run_fiftyone.py
@@ -6,14 +6,9 @@ from tqdm import tqdm
 Runs fiftyone. If a dataset was not yet created (or if param delete=False), it
 loads a dataset (reads paths from common.py), adds tags to the samples (dataset
 name) and then runs fiftyone.
-
-Remember to convert updated data to COCO format before rebuilding the dataset")
 """
 
-
 def run_fiftyone(delete=False):
-
-    print("Remember to convert updated data to COCO format before rebuilding the dataset")
 
     dataset_name = "dataset"
 
@@ -21,26 +16,31 @@ def run_fiftyone(delete=False):
         dataset = fo.load_dataset(dataset_name) 
 
     else:
+        print("Creating dataset")
         dataset = fo.Dataset.from_dir(
             name=dataset_name,
-            data_path=common.datasets_path,
-            labels_path=common.dataset_coco_filepath,
+            data_path=common.datasets_dirpath,
+            labels_path=common.dataset_filepath,
             dataset_type=fo.types.COCODetectionDataset
         )
 
         print("Updating tags")
         for sample in tqdm(dataset):
-            ds_name = sample["filepath"].split(common.datasets_path)[1].split("/")[0]
-            if ds_name not in sample.tags:
-                sample.tags.append(ds_name)
-                sample.save()
+            ds_name = sample["filepath"].split(common.datasets_dirpath)[1].split("/")[0]
+            sample.tags.append(ds_name)
+            sample.save()
 
     if not dataset.persistent:
         dataset.persistent = True
 
     if delete:
-        dataset.delete()
-        run_fiftyone()
+        s = input("Are you sure you want to delete the existing dataset? (y/n): ")
+        if s == "y":
+            dataset.delete()
+            print("Dataset deleted")
+            run_fiftyone()
+        else:
+            print("Deletion aborted. Exiting")
         return
 
     view = dataset.view()
@@ -48,7 +48,6 @@ def run_fiftyone(delete=False):
 
     session = fo.launch_app(dataset)
     session.wait()
-
 
 if __name__ == "__main__":
     run_fiftyone(delete=False)
